requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
flask-sqlalchemy>=3.0.0
sqlalchemy>=1.4.0
cryptography>=3.0.0
bcrypt>=3.0.0
python-dotenv>=0.19.0
pyjwt>=2.0.0
python-dateutil>=2.8.0
click>=8.0.0